refactor: move into admin.tsx
